+++ modules/template/widgets/TemplatePageEditButton.php
@@ -9,9 +9,8 @@
 namespace humhub\modules\custom_pages\modules\template\widgets;
 
 use humhub\components\Widget;
-use Yii;
+use humhub\modules\content\helpers\ContentContainerHelper;
 use humhub\modules\custom_pages\models\CustomPage;
-use humhub\modules\custom_pages\modules\template\models\TemplateInstance;
 
 /**
  * User Administration Menu
@@ -43,14 +42,9 @@ class TemplatePageEditButton extends Widget
      */
     public function run()
     {
-        $space = Yii::$app->controller->contentContainer ?? null;
-
-        $templateInstance = TemplateInstance::findOne(['page_id' => $this->page->id]);
-
         return $this->render('templatePageEditButton', [
-            'sguid' => $space ? $space->guid : null,
-            'pageId' => $this->page->id,
-            'templateInstance' => $templateInstance,
+            'page' => $this->page,
+            'container' => ContentContainerHelper::getCurrent(),
         ]);
     }
 }