+++ views/container/markdown.php
@@ -5,11 +5,11 @@
  * @license https://www.humhub.com/licences
  */
 
+use humhub\components\View;
 use humhub\modules\content\widgets\richtext\RichText;
-use humhub\modules\custom_pages\assets\InlineStyleAssets;
+use humhub\modules\custom_pages\assets\Assets;
 use humhub\modules\custom_pages\models\CustomPage;
 use humhub\modules\custom_pages\modules\template\widgets\PageConfigurationButton;
-use humhub\modules\ui\view\components\View;
 use yii\helpers\Html;
 
 /* @var $this View */
@@ -18,7 +18,7 @@ use yii\helpers\Html;
 
 $cssClass = ($page->hasAttribute('cssClass') && !empty($page->cssClass)) ? $page->cssClass : 'custom-pages-page';
 
-InlineStyleAssets::register($this);
+Assets::register($this);
 ?>
 <?= PageConfigurationButton::widget() ?>
 <div class="panel panel-default <?= Html::encode($cssClass) ?>">