+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 1.11.0-beta.8 (Unreleased)
 -----------------------------
 - Enh #432: Add Yii formatter for Twig filters
+- Enh #437: Add context menu to Structure View
 
 1.11.0-beta.7 (July 4, 2025)
 -----------------------------